refactor(rtc_tenclave): import and use C NaCl Box API zero-padding constants
This makes it a bit easier to see what padding is derived from where,
and to check the implementation against the calling convention docs.

Note that some of the generic const parameters can't trivially be
substituted with these constants (or expressions involving them), so I
added inline comments for those, for now.